SCRUM-40: -name change on transitions-class to petri-net-transitions -name property from PN-Interface moved to Transition-class
